moved s3 file to access function
